+++ src/main/bitbucket/PullRequestFormatter.h
@@ -2,6 +2,8 @@
 #define MUSTARD_PULLREQUESTFORMATTER_H
 
 #include <string>
+#include <vector>
+#include<functional>
 #include "model/PullRequest.h"
 
 using namespace std;
@@ -9,6 +11,7 @@ class PullRequestFormatter
 {
 public:
     static string format ( const PullRequest &pullRequest );
+    static string shortFormat ( const vector<PullRequest> &pullRequest, function<bool ( const PullRequest & ) > highlight );
 
 private:
     static ostream &formatBold ( ostream &stream );