+++ src/CHANGELOG.md
@@ -1,2 +1,3 @@
 ## Changelog for mustard-cli
+- feature: *mustard inbox* can display all your currently pending PRs
 - fix: libsecret would not ask for password in case of a ["cold keyring"](https://gitlab.gnome.org/GNOME/libsecret/issues/7)